refactor method to api project.
