+++ ThaiOpenBraille.Api/BrailleTabe.cs
@@ -1,0 +1,165 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace ThaiOpenBraille.Api
+{
+    public class BrailleTable
+    {
+        public Dictionary<String, String> AllBrailleTable = new Dictionary<String, String>() {
+
+		   {" ",'\u2800'.ToString()},
+
+            // English
+            {"a","⠁"},
+            {"b","⠃"},
+            {"c","⠉"},
+            {"d","⠙"},
+            {"e","⠑"},
+            {"f","⠋"},
+            {"g","⠛"},
+            {"h","⠓"},
+            {"i","⠊"},
+            {"j","⠚"},
+            {"k","⠅"},
+            {"l","⠇"},
+            {"m","⠍"},
+            {"n","⠝"},
+            {"o","⠕"},
+            {"p","⠏"},
+            {"q","⠟"},
+            {"r","⠗"},
+            {"s","⠎"},
+            {"t","⠞"},
+            {"u","⠥"},
+            {"v","⠧"},
+            {"w","⠺"},
+            {"x","⠭"},
+            {"y","⠽"},
+            {"z","⠵"},
+
+            //Numbers
+            {"0", "⠚"},
+            {"1", "⠁"},
+            {"2", "⠃"},
+            {"3", "⠉"},
+            {"4", "⠙"},
+            {"5", "⠑"},
+            {"6", "⠋"},
+            {"7", "⠛"},
+            {"8", "⠓"},
+            {"9", "⠊"},
+
+            //พยัญชนะ
+            {"ก", "⠛"},
+            {"ข", "⠅"},
+            {"ฃ", "⠴⠅"},
+            {"ค", "⠥"},
+            {"ฅ", "⠤⠥"},
+            {"ฆ", "⠠⠥"},
+            {"ง", "⠻"},
+            {"จ", "⠚"},
+            {"ฉ", "⠌"},
+            {"ช", "⠬"},
+            {"ซ", "⠮"},
+            {"ฌ", "⠠⠬"},
+            {"ญ", "⠠⠽"},
+            {"ฎ", "⠠⠙"},
+            {"ฏ", "⠠⠳"},
+            {"ฐ", "⠠⠞"},
+            {"ฑ", "⠠⠾"},
+            {"ฒ", "⠤⠾"},
+            {"ณ", "⠠⠝"},
+            {"ด", "⠙"},
+            {"ต", "⠳"},
+            {"ถ", "⠞"},
+            {"ท", "⠾"},
+            {"ธ", "⠴⠾"},
+            {"น", "⠝"},
+            {"บ", "⠧"},
+            {"ป", "⠯"},
+            {"ผ", "⠏"},
+            {"ฝ", "⠭"},
+            {"พ", "⠹"},
+            {"ฟ", "⠫"},
+            {"ภ", "⠠⠹"},
+            {"ม", "⠍"},
+            {"ย", "⠽"},
+            {"ร", "⠗"},
+            {"ล", "⠇"},
+            {"ว", "⠺"},
+            {"ศ", "⠠⠎"},
+            {"ษ", "⠤⠎"},
+            {"ส", "⠎"},
+            {"ห", "⠓"},
+            {"ฬ", "⠠⠇"},
+            {"อ", "⠕"},
+            {"ฮ", "⠿"},
+            {"ฤ", "⠗⠂"},
+            {"ฦ", "⠇⠂"},
+
+
+
+            //สระ
+            {"ะ", "⠁"},
+            {"ั", "⠜"},
+            {"า", "⠡"},
+            {"ำ", "⠵"},
+            {"ิ", "⠃"},
+            {"ี", "⠆"},
+            {"ึ", "⠪"},
+            {"ื", "⠢"},
+            {"ุ", "⠉"},
+            {"ู", "⠒"},
+            {"เ", "⠋"},
+            {"แ", "⠣"},
+            {"โ", "⠊"},
+            {"ใ", "⠱⠂"},
+            {"ไ", "⠱"},
+            {"ๆ", "⠂"},
+            {"็", "⠄"},
+            {"่", "⠔"},
+            {"้", "⠲"},
+            {"๊", "⠶"},
+            {"๋", "⠦"},
+            {"์", "⠴"},
+            {"เอาะ", "⠕⠁"},
+            {"เออ", "⠩"},
+            {"เอา", "⠖"},
+            {"อัว", "⠑"},
+            {"เอีย", "⠷"},
+            {"เอือ", "⠟"},
+            {"ฯลฯ", "⠰⠇"},
+            {"ฯ", "⠰⠆"},
+            {"ฺ", "⠄"},
+            {"ํ", "⠐"},
+            {".", "⠲"},
+            {"เทอญ", "⠋⠾⠕⠠⠽"},
+            {"เทอม", "⠋⠾⠕⠍"},
+
+            //TH
+            {"T,", "⠸⠂"},
+            {"T?", "⠸⠦"},
+            {"T!", "⠸⠖"},
+            {"T;", "⠸⠆"},
+            {"T:", "⠸⠒"},
+            {"T)", "⠾"},
+            {"T(", "⠷"},
+
+            //EN
+            {"E,", "⠂"},
+            {"E?", "⠦"},
+            {"E!", "⠖"},
+            {"E;", "⠆"},
+            {"E:", "⠒"},
+            {"E)", "⠾"},
+            {"E(", "⠷"},
+
+            {"-", "⠤"},
+            {"=", "⠒⠒"},
+            {"/", "⠸⠌"},
+            {"\\", "⠸⠡"},
+            {"%", "⠨⠴"},
+            {"$", "⠐⠎"},
+        };
+    }
+}
