+++ ThaiOpenBraille.Api/IWordManager.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ThaiOpenBraille.Api
+{
+    public interface IWordManager
+    {
+		string Output();
+	}
+}
